registration page

+++ note/views.py
@@ -1,9 +1,37 @@
 from django.shortcuts import render,redirect
 from django.contrib import messages
+from django.contrib.auth.forms import UserCreationForm
+from django.contrib.auth import login,authenticate
 from .models import notes
 from django.http import HttpResponse
-from .forms import notesupdateform
+from .forms import notesupdateform,createuserform
 # Create your views here.
+def registration(request):
+  form = createuserform()
+  if request.method == 'POST':
+    form = createuserform(request.POST)
+    if form.is_valid():
+      form.save()
+      #user = form.cleaned_data.get(username)
+      messages.success(request,'Account is created ')
+      return redirect('login')
+  context={'form':form}
+  return render(request,'note/register.html',context)
+
+def loginpage(request):
+  if request.method == 'POST':
+    print(request.POST)
+    username=request.POST.get('username')
+    password=request.POST.get('password')
+    user = authenticate(request,username=username,password=password)
+    if user is not None:
+      login(request,user)
+      return redirect('home') 
+    else:
+      messages.info(request,'Username or Password incorrect')
+  context={}
+  return render(request,'note/login.html',context)
+
 def index(request):
   context ={'notes':notes.objects.all(),}
   return render(request,'note/home.html',context)
@@ -16,7 +44,7 @@ def add(request):
     form = notesupdateform(request.POST,)
     if form.is_valid():
       form.save()
-      messages.success(request,f'Notes succesfully created')
+      messages.success(request,'Notes succesfully created')
       return redirect('home')
   else:
     form = notesupdateform()
@@ -24,7 +52,8 @@ def add(request):
   return render(request,'note/add.html',context)
 
 def delete(request,note_id):
-  ele = notes.objects.get(id=note_id)
-  ele.delete()
-  messages.info(request,"Notes Removed")
-  return redirect('home')
+  if request.method == 'POST':
+    ele = notes.objects.get(id=note_id)
+    ele.delete()
+    messages.info(request,"Notes Removed")
+    return redirect('home')
